feat(api): add `range`, `count`, `reg` to `nvim_parse_cmd`
